+++ ui/src/api/variables/mutate/create.ts
